Move Ruler's inner classes out. * Possibly make a package for them in the future.
