+++ src/ispd/gui/iconico/RulerOrientation.java
@@ -1,0 +1,10 @@
+package ispd.gui.iconico;
+
+/**
+ * {@link RulerOrientation} is an enumeration that stores the
+ * available ruler orientation.
+ */
+public enum RulerOrientation {
+    HORIZONTAL,
+    VERTICAL
+}
